requirements.txt
environs==9.5.0
flask==2.1.2
flask-sqlalchemy==2.5.1
flask-migrate==3.1.0
flask-marshmallow==0.14.0
marshmallow-sqlalchemy==0.28.0
pymysql==1.0.2
flasgger==0.9.5
gunicorn==20.1.0